Pre-Formatting IO Mismatch check removal
Leave checking and handling of IO mismatches to the component using a MarkupLanguage's methods as the desired functionality will likely vary and mismatches won't break the default approach. Additionally, the developer should be responsible for identifying and eliminating any undesired mismatches as opposed to an administrator being required to fix those that slip through or ignore those that are intentional.